applied multer-s3, installed multer-s3, aws-sdk and modified multer middleware and completely modified pets/register,edit,remove except comments/register

+++ controller/pets/remove.js
@@ -1,28 +1,41 @@
 const { Pets, PetsImages } = require("../../models");
-const fs = require("fs");
-const path = require("path");
+const { s3 } = require("../../routes/multer");
 
 module.exports = async (req, res) => {
+  const { id } = req.user;
+
+  const compareUser = await Pets.findOne({
+    where: { id: req.params.id },
+  });
+  if (compareUser.userId !== id) {
+    return res.status(400).json({ message: "not Authorized" });
+  }
+
   const findImages = await PetsImages.findAll({
     where: { petId: req.params.id },
     attributes: ["imagePath"],
   });
 
-  for (let i = 0; i < findImages.length; i += 1) {
-    if (findImages[i].imagePath.includes("uploads")) {
-      fs.unlink(
-        path.join(__dirname, "../..", findImages[i].imagePath),
-        (err) => {
-          if (err) throw err;
-        }
-      );
+  if (findImages.length) {
+    const fileUrls = [];
+    for (let i = 0; i < findImages.length; i += 1) {
+      const fileUrl = findImages[i].imagePath.split("/");
+      const delFileName = fileUrl[fileUrl.length - 1];
+      fileUrls.push({ Key: delFileName });
     }
+
+    const params = {
+      Bucket: "missing-animals-images",
+      Delete: { Objects: fileUrls },
+    };
+    s3.deleteObjects(params, (err) => {
+      if (err) return next(err);
+    });
+
+    await PetsImages.destroy({ where: { petId: req.params.id }, force: true });
   }
 
   await Pets.destroy({ where: { id: req.params.id }, force: true });
-  await PetsImages.destroy({ where: { petId: req.params.id }, force: true });
 
   res.status(200).json({ message: "remove OK" });
-
-  // console.log(__dirname); --> /Users/mac/Desktop/missing-animals/Missing_Animal_Server/controller/pets
 };
